api/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
python-dotenv>=1.0,<2.0
mysqlclient>=2.2,<3.0

